+++ frontend/src/useRecipes.ts
@@ -26,7 +26,5 @@ export default function useRecipes() {
             .catch(() => console.error("post on /api/recipes not successful"))
     }
 
-
-
     return {recipes, addRecipe}
 }
